Move get_run_analysis into .analysis.free_energy
