+++ covid_moonshot/analysis/free_energy.py
@@ -4,7 +4,14 @@ from typing import List, Optional, Tuple
 import numpy as np
 from pymbar import BAR
 from pymbar.mbar import MBAR
-from covid_moonshot.core import FreeEnergy, GenAnalysis, PhaseAnalysis, Work
+from covid_moonshot.core import (
+    Binding,
+    FreeEnergy,
+    GenAnalysis,
+    PhaseAnalysis,
+    RunAnalysis,
+    Work,
+)
 
 
 class InsufficientDataError(ValueError):
@@ -247,3 +254,30 @@ def get_phase_analysis(
     free_energy = get_free_energy(ws, min_num_work_values=min_num_work_values)
 
     return PhaseAnalysis(free_energy=free_energy, gens=gens)
+
+
+def get_run_analysis(
+    run: int, complex_works: List[Work], solvent_works: List[Work],
+) -> RunAnalysis:
+
+    try:
+        complex_phase = get_phase_analysis(run, "complex", complex_works)
+    except ValueError as e:
+        raise ValueError(f"Failed to analyze complex: {e}")
+
+    try:
+        solvent_phase = get_phase_analysis(run, "solvent", solvent_works)
+    except ValueError as e:
+        raise ValueError(f"Failed to analyze solvent: {e}")
+
+    binding = Binding(
+        delta_f=solvent_phase.free_energy.delta_f - complex_phase.free_energy.delta_f,
+        ddelta_f=np.sqrt(
+            complex_phase.free_energy.ddelta_f ** 2
+            + solvent_phase.free_energy.ddelta_f ** 2
+        ),
+    )
+
+    return RunAnalysis(
+        complex_phase=complex_phase, solvent_phase=solvent_phase, binding=binding
+    )
